+++ src/components/Footer.jsx
@@ -1,0 +1,10 @@
+export const Footer = () => {
+  return (
+    <>
+      <footer>
+        <p>2024 by Arnau Vidal </p>
+        <img src="./src/assets/technigo-logo.svg" alt="Technigo logo" className="logo" />
+      </footer>
+    </>
+  )
+}
